db updated

+++ model/da/db.sql
@@ -42,7 +42,7 @@ group by person_id;
 create table FinancialDoc_tbl
 (
     id          int primary key auto_increment,
-    amount      decimal(15, 2) not null,
+    amount      int not null,
     date_time   datetime       not null,
     doc_type    nvarchar(8)    not null,
     description nvarchar(100)  not null,
